+++ ext/log-processor/src/test/java/core/log/job/CleanupOldIndexJobTest.java
@@ -30,7 +30,11 @@ class CleanupOldIndexJobTest {
         job = new CleanupOldIndexJob();
         job.elasticSearch = elasticSearch;
         job.indexService = new IndexService();
-        job.jobConfig = defaultJobConfig();
+
+        var jobConfig = new JobConfig();
+        jobConfig.indexRetentionDays = 30;
+        jobConfig.indexOpenDays = 7;
+        job.jobConfig = jobConfig;
     }
 
     @Test
@@ -48,36 +52,9 @@ class CleanupOldIndexJobTest {
         verify(elasticSearch).closeIndex("action-2017.11.01");
     }
 
-    @Test
-    void customJobConfig() {
-        var jobConfig = new JobConfig();
-        jobConfig.indexAliveDays = 7;
-        jobConfig.indexOpenDays = 3;
-        job.jobConfig = jobConfig;
-        var state = new ClusterStateResponse();
-        state.metadata = new ClusterStateResponse.Metadata();
-        state.metadata.indices = Map.of("action-2017.10.29", index(ClusterStateResponse.IndexState.OPEN),
-                "action-2017.10.30", index(ClusterStateResponse.IndexState.CLOSE),
-                "action-2017.11.05", index(ClusterStateResponse.IndexState.OPEN));
-        when(elasticSearch.state()).thenReturn(state);
-
-        job.cleanup(LocalDate.of(2017, 11, 8));
-
-        verify(elasticSearch).deleteIndex("action-2017.10.29");
-        verify(elasticSearch).deleteIndex("action-2017.10.30");
-        verify(elasticSearch).closeIndex("action-2017.11.05");
-    }
-
     private ClusterStateResponse.Index index(ClusterStateResponse.IndexState state) {
         ClusterStateResponse.Index index = new ClusterStateResponse.Index();
         index.state = state;
         return index;
-    }
-
-    private JobConfig defaultJobConfig() {
-        var jobConfig = new JobConfig();
-        jobConfig.indexAliveDays = 30;
-        jobConfig.indexOpenDays = 7;
-        return jobConfig;
     }
 }